unittest
least stupid solution chosen? :)

the idea is that you can test public APIs in unittest.c. and if you
want to test private stuff just make it public (or optionally have a
decorator #define that resolves to "static" except in unittest?)

I considered the "unit test included in .c-file" style. but
unfortunately it doesn't really work well with the chibicc heritage
because there are many cross references between units. I considered
using a linker-hack that ignores unresolved symbols. but that's
rather ugly, non-portable, and linker errors are "demoted" to
runtime errors (also I couldn't easily get it to work).

I also considered porting it to a stb-style all-in-one-file library
but it would exceed 10ksloc. it would also obfuscate the chibicc
heritage which makes it harder to do things like backporting features
to chibicc or figuring out what has changed. I also considered driving
the code towards a "leaf style" where most files don't have cross
references, but that just feels like a "lite version" of stb-style...

another reason for the chosen solution is that wasm exports have to be
marked with:
   __attribute__((visibility("default")))
so making private functions public to be able to test them isn't that
bad.